refactor: theme is now kaamiki
