+++ kaamiki/base/static/base.js
