+++ kaamiki/base/static/sphinx-design.css
