+++ kaamiki/base/templates/previous_next_pages.html.jinja
