+++ kaamiki/extensions/__init__.py
@@ -4,9 +4,9 @@ Theme Extension Manager
 
 Author: Akshay Mestry <xa@mes3.dev>
 Created on: 22 February, 2025
-Last updated on: 02 November, 2025
+Last updated on: 26 April, 2026
 
-This module manages this theme's custom directive and roles.
+This module manages Kaamiki Sphinx Theme's custom directive and roles.
 
 .. deprecated:: 19.10.2025
